+++ crates/dies-executor/src/roles/mod.rs
@@ -1,0 +1,3 @@
+mod goalkeeper;
+mod passer;
+mod receiver;
